Revert "Move GraphQL API component to fuel-indexer-api-server (#257)"
This reverts commit 2ce05f696a3bd01cd945b0c76450091d97f122bb.
