+++ fuel-indexer-api-server/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [lib]
-name = "fuel_indexer_api_server"
+name = "api_server"
 path = "src/lib.rs"
 
 [[bin]]
@@ -19,13 +19,9 @@ path = "src/main.rs"
 anyhow = "1.0"
 async-std = "1.10.0"
 axum = "0.4"
-fuel-indexer-database = { path =  "../fuel-indexer-database" }
-fuel-indexer-database-types = { path = "../fuel-indexer-database/database-types" }
+fuel-indexer = { path = "../fuel-indexer" }
 fuel-indexer-lib = { path = "../fuel-indexer-lib" }
 fuel-indexer-schema = { path = "../fuel-indexer-schema", features = ["db-models"] }
-http = "0.2"
-hyper = { version = "0.14", features = ["client", "http2", "http1", "runtime" ]}
-hyper-tls = "0.5"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0", features = ["raw_value"] }
 serde_yaml = "0.8"
