+++ fuel-indexer-api-server/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod api;
+
