+++ fuel-indexer-api-server/src/main.rs
@@ -1,6 +1,8 @@
 use anyhow::Result;
-use fuel_indexer_api_server::api::GraphQlApi;
-use fuel_indexer_lib::config::{IndexerArgs, IndexerConfig, Parser};
+use fuel_indexer::{
+    config::{IndexerArgs, Parser},
+    GraphQlApi, IndexerConfig,
+};
 use tracing::info;
 use tracing_subscriber::filter::EnvFilter;
