+++ fuel-indexer-lib/Cargo.toml
@@ -8,10 +8,7 @@ edition = "2021"
 
 [dependencies]
 anyhow = { version = "1.0", default-features = false }
-async-std = "1.10.0"
-clap = { version = "3.1", features = ["cargo", "derive", "env"] }
 serde = { version = "1.0", features = ["derive"] }
-serde_yaml = "0.8"
 sha2 = "0.9"
 tokio = { version = "1.8", features = ["time"] }
 tracing = "0.1"
