+++ fuel-indexer-schema/src/db/tables.rs
@@ -1,8 +1,6 @@
 use crate::db::models::{ColumnIndex, CreateStatement, ForeignKey, IdCol, IndexMethod};
-use crate::{get_schema_types, schema_version, type_id, BASE_SCHEMA};
-use fuel_indexer_database::{
-    queries, DbType, IndexerConnection, IndexerConnectionPool, IndexerDatabaseError,
-};
+use crate::{get_schema_types, type_id, BASE_SCHEMA};
+use fuel_indexer_database::{queries, DbType, IndexerConnection, IndexerConnectionPool};
 use fuel_indexer_database_types::*;
 use graphql_parser::parse_schema;
 use graphql_parser::schema::{Definition, Field, SchemaDefinition, Type, TypeDefinition};
@@ -326,43 +324,6 @@ impl SchemaBuilder {
             }
             o => panic!("Got a non-object type! {:?}", o),
         }
-    }
-}
-
-/// Responsible for laying down graph schemas, processes schema upgrades.
-pub struct SchemaManager {
-    pool: IndexerConnectionPool,
-}
-
-impl SchemaManager {
-    pub async fn new(
-        db_conn: impl Into<String>,
-    ) -> Result<SchemaManager, IndexerDatabaseError> {
-        let pool = IndexerConnectionPool::connect(&db_conn.into()).await?;
-
-        Ok(SchemaManager { pool })
-    }
-
-    pub async fn new_schema(&self, name: &str, schema: &str) -> Result<(), sqlx::Error> {
-        let mut connection = self.pool.acquire().await?;
-
-        // TODO: Not doing much with version, but might be useful if we
-        //       do graph schema upgrades
-        let version = schema_version(schema);
-
-        if !queries::schema_exists(&mut connection, name, &version).await? {
-            let _db_schema =
-                SchemaBuilder::new(name, &version, self.pool.database_type())
-                    .build(schema)
-                    .commit_metadata(&mut connection)
-                    .await?;
-        }
-        Ok(())
-    }
-
-    pub async fn load_schema_wasm(&self, name: &str) -> Result<Schema, sqlx::Error> {
-        // TODO: might be nice to cache this data in server?
-        Schema::load_from_db(&self.pool, name).await
     }
 }
