+++ fuel-indexer-tests/tests/executor.rs
@@ -1,6 +1,6 @@
-use fuel_indexer::{executor::WasmIndexExecutor, Executor, IndexerError};
-use fuel_indexer_lib::manifest::Manifest;
-use fuel_indexer_schema::db::tables::SchemaManager;
+use fuel_indexer::{
+    executor::WasmIndexExecutor, Executor, IndexerError, Manifest, SchemaManager,
+};
 use fuel_indexer_schema::BlockData;
 use fuel_indexer_tests::assets::{BAD_MANIFEST, BAD_WASM_BYTES, MANIFEST, WASM_BYTES};
 use fuel_tx::Receipt;
