+++ fuel-indexer/Cargo.toml
@@ -13,9 +13,9 @@ async-trait = "0.1"
 axum = { version = "0.4", features = ["multipart"] }
 bincode = "1.3.3"
 cfg-if = "1.0"
+clap = { version = "3.1", features = ["cargo", "derive", "env"] }
 fuel-core = { version = "0.10", default-features = false }
 fuel-gql-client = { version = "0.10" }
-fuel-indexer-api-server = { path = "../fuel-indexer-api-server" }
 fuel-indexer-database = { path =  "../fuel-indexer-database" }
 fuel-indexer-database-types = { path = "../fuel-indexer-database/database-types" }
 fuel-indexer-lib = { path = "../fuel-indexer-lib" }
