+++ fuel-indexer/src/manifest.rs
